requirements.txt
contourpy>=1.3.3
cycler>=0.12.1
filelock>=3.19.1
fonttools>=4.59.1
fsspec>=2025.7.0
Jinja2>=3.1.6
joblib>=1.5.1
kiwisolver>=1.4.9
MarkupSafe>=3.0.2
matplotlib>=3.10.5
mpmath>=1.3.0
networkx>=3.5
numpy>=2.3.2
packaging>=25.0
pillow>=11.3.0
pyparsing>=3.2.3
python-dateutil>=2.9.0.post0
PyYAML>=6.0.2
scikit-learn>=1.7.1
scipy>=1.16.1
six>=1.17.0
sympy>=1.14.0
threadpoolctl>=3.6.0
torch>=2.8.0
typing_extensions>=4.14.1
